+++ ailes/Social_networks/templates/Social_networks/signup.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Document</title>
-    <link rel="stylesheet" type="text/css" href="../../static/App/css/style.css">
+    <link rel="stylesheet" type="text/css" href="../../static/Social_networks_app/css/style.css">
 </head>
 <body>
     <form action="action_page.php">
