+++ ailes/ailes/urls.py
@@ -19,5 +19,5 @@ from django.urls import path, include
 
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path("", include("App.urls"))
+    path("", include("Social_networks.urls"))
 ]
